refactor: update item service and controller to use object destructuring and improve error handling

+++ src/modules/items/items.microservice.controller.ts
@@ -16,22 +16,47 @@ export class ItemsMicroserviceController {
   }
 
   @MessagePattern({ cmd: ItemMessagePattern.FIND_ONE })
-  async getItem(id: string): Promise<ItemEntity> {
+  async getItem({ id }: { id: string }): Promise<ItemEntity> {
+    console.log('getItem. id: ', id);
+    if (!id) {
+      throw new Error(`Microservice - getItem. Invalid id: ${id}`);
+    }
     return this.itemsService.findOne(id);
   }
 
   @MessagePattern({ cmd: ItemMessagePattern.CREATE })
-  async create(dto: CreateItemDto): Promise<ItemEntity> {
+  async create({ dto }: { dto: CreateItemDto }): Promise<ItemEntity> {
+    console.log('create. dto: ', dto);
+    if (!dto) {
+      throw new Error(
+        `Microservice - create. Invalid create payload: ${JSON.stringify({ dto })}`,
+      );
+    }
     return this.itemsService.create(dto);
   }
 
   @MessagePattern({ cmd: ItemMessagePattern.UPDATE })
-  async update(id: string, dto: UpdateItemDto): Promise<ItemEntity> {
+  async update({
+    id,
+    dto,
+  }: {
+    id: string;
+    dto: UpdateItemDto;
+  }): Promise<ItemEntity> {
+    if (!id || !dto) {
+      throw new Error(
+        `Microservice - update. Invalid update payload: ${JSON.stringify({ id, dto })}`,
+      );
+    }
     return this.itemsService.update(id, dto);
   }
 
   @MessagePattern({ cmd: ItemMessagePattern.DELETE })
-  async delete(id: string): Promise<void> {
-    return this.itemsService.delete(id);
+  async delete({ id }: { id: string }): Promise<{ success: boolean }> {
+    if (!id) {
+      throw new Error(`Microservice - delete. Invalid id: ${id}`);
+    }
+    await this.itemsService.delete(id);
+    return { success: true };
   }
 }
